Add WithSkia component
